Add files via upload
A. Write a program to determine grades in a course with three quizzes, each scored on a basis of 10 points. The final
letter grade is determined according to the following rules:
o A is 90 or over;
o B is 80 or over;
o C is 70 or over;
o D is 60 or over;
o F is below 60.