Create HEIGHT
Write a program that guesses the user’s height. Assume minimum height is 100cm and maximum height is 200cm. The
program makes a first guess and then asks the user if it is too high, too low, or correct. The program continues to guess
and ask the user for feedback until the user says the guess is correct or until three tries have been made, whichever comes
first.